" likes/bookmarks to constants, fixed user name "
