+++ constants/dummyBookmarks.ts
@@ -1,0 +1,22 @@
+const dummyBookmarks = [
+  {
+    id: 1,
+    title: "Title 1",
+    description: "Description for Bookmark 1",
+    url: "#",
+  },
+  {
+    id: 2,
+    title: "Title 2",
+    description: "Description for Bookmark 2",
+    url: "#",
+  },
+  {
+    id: 3,
+    title: "Title 3",
+    description: "Description for Bookmark 3",
+    url: "#",
+  },
+]
+
+export default dummyBookmarks
