+++ constants/dummyLikes.ts
@@ -1,0 +1,22 @@
+const dummyLikes = [
+  {
+    id: 1,
+    title: "Title 1",
+    description: "Description for Likes 1",
+    url: "#",
+  },
+  {
+    id: 2,
+    title: "Title 2",
+    description: "Description for Likes 2",
+    url: "#",
+  },
+  {
+    id: 3,
+    title: "Title 3",
+    description: "Description for Likes 3",
+    url: "#",
+  },
+]
+
+export default dummyLikes
