+++ views/profil/BookmarkSection/Bookmark.tsx
@@ -1,26 +1,6 @@
 import Image from "next/image"
 import Link from "next/link"
-
-const dummyBookmarks = [
-  {
-    id: 1,
-    title: "Title 1",
-    description: "Description for Bookmark 1",
-    url: "#",
-  },
-  {
-    id: 2,
-    title: "Title 2",
-    description: "Description for Bookmark 2",
-    url: "#",
-  },
-  {
-    id: 3,
-    title: "Title 3",
-    description: "Description for Bookmark 3",
-    url: "#",
-  },
-]
+import dummyBookmarks from "@constants/dummyBookmarks"
 
 const BookmarkSection = () => {
   return (
