+++ views/profil/LikedSection/LikeSection.tsx
@@ -1,26 +1,6 @@
 import Image from "next/image"
 import Link from "next/link"
-
-const dummyLikes = [
-  {
-    id: 1,
-    title: "Title 1",
-    description: "Description for Likes 1",
-    url: "#",
-  },
-  {
-    id: 2,
-    title: "Title 2",
-    description: "Description for Likes 2",
-    url: "#",
-  },
-  {
-    id: 3,
-    title: "Title 3",
-    description: "Description for Likes 3",
-    url: "#",
-  },
-]
+import dummyLikes from "@constants/dummyLikes"
 
 const LikedSection = () => {
   return (
